requirements.txt
pandas>=1.5.0
requests>=2.28.0
pyyaml>=6.0
jupyter>=1.0.0
numpy>=1.24.0
python-dotenv>=0.19.0
matplotlib>=3.6.0
seaborn>=0.11.0
openpyxl>=3.0.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.0

# Documentation server (simple HTTP) & MkDocs site
markdown>=3.4.0
mkdocs>=1.5.0
mkdocs-material>=9.5.0
pymdown-extensions>=10.0
mkdocs-mermaid2-plugin>=1.2.0